move useSongFileDuplicates()
