+++ projects/graphql-client/src/index.ts
@@ -43,3 +43,4 @@ export * from "./remote/mutations/useUpdateAuth"
 export * from "./remote/mutations/useLogout"
 export * from "./remote/mutations/useIssueAuthToken"
 export * from "./remote/mutations/useRestorePassword"
+export * from "./remote/mutations/useSongFileDuplicates"
